% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dreams_geographies}
\alias{dreams_geographies}
\title{DREAMS Country names, DSNU names, ISO2, and ISO3 codes}
\format{
A data frame with 217 rows and 7 variables:
\describe{
  \item{ou_uid}{uid for DSNU}
  \item{dsnu}{DREAMS district name}
  \item{ou}{Country name}
  \item{snu1}{SNU1 name}
  \item{iso3}{Three-char country code}
  \item{iso2}{Two-char country code}
  \item{iso2_naomi}{Three-char country code for NAOMI}
}
}
\usage{
dreams_geographies
}
\description{
DREAMS Country names, DSNU names, ISO2, and ISO3 codes
}
\keyword{datasets}
