% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dreams_iso}
\alias{dreams_iso}
\title{ISO3 and ISO2 Codes for DREAMS Countries}
\format{
A data frame with 16 rows and 3 variables:
\describe{
  \item{country}{Country name}
  \item{iso3}{Three-char country code}
  \item{iso2}{Two-char country code}
}
}
\usage{
dreams_iso
}
\description{
ISO3 and ISO2 Codes for DREAMS Countries
}
\keyword{datasets}
